resolve pylint error
